Fix MSVC build

+++ source/MRMesh/MRZlib.cpp
@@ -64,20 +64,20 @@ VoidOrErrStr zlibCompressStream( std::istream& in, std::ostream& out, int level 
         if ( in.bad() )
             return unexpected( "I/O error" );
         stream.next_in = reinterpret_cast<uint8_t*>( inChunk.data() );
-        stream.avail_in = (uint32_t)in.gcount();
-        assert( stream.avail_in <= inChunk.size() );
+        stream.avail_in = (unsigned)in.gcount();
+        assert( stream.avail_in <= (unsigned)inChunk.size() );
 
         const auto flush = in.eof() ? Z_FINISH : Z_NO_FLUSH;
         do
         {
             stream.next_out = reinterpret_cast<uint8_t*>( outChunk.data() );
-            stream.avail_out = outChunk.size();
+            stream.avail_out = (unsigned)outChunk.size();
             ret = deflate( &stream, flush );
             if ( Z_OK != ret && Z_STREAM_END != ret )
                 return unexpected( zlibToString( ret ) );
 
-            assert( stream.avail_out <= outChunk.size() );
-            out.write( outChunk.data(), outChunk.size() - stream.avail_out );
+            assert( stream.avail_out <= (unsigned)outChunk.size() );
+            out.write( outChunk.data(), (unsigned)outChunk.size() - stream.avail_out );
             if ( out.bad() )
                 return unexpected( "I/O error" );
         }
@@ -109,19 +109,19 @@ VoidOrErrStr zlibDecompressStream( std::istream& in, std::ostream& out )
         if ( in.bad() )
             return unexpected( "I/O error" );
         stream.next_in = reinterpret_cast<uint8_t*>( inChunk.data() );
-        stream.avail_in = (uint32_t)in.gcount();
-        assert( stream.avail_in <= inChunk.size() );
+        stream.avail_in = (unsigned)in.gcount();
+        assert( stream.avail_in <= (unsigned)inChunk.size() );
 
         do
         {
             stream.next_out = reinterpret_cast<uint8_t*>( outChunk.data() );
-            stream.avail_out = outChunk.size();
+            stream.avail_out = (unsigned)outChunk.size();
             ret = inflate( &stream, Z_NO_FLUSH );
             if ( Z_OK != ret && Z_STREAM_END != ret )
                 return unexpected( zlibToString( ret ) );
 
-            assert( stream.avail_out <= outChunk.size() );
-            out.write( outChunk.data(), outChunk.size() - stream.avail_out );
+            assert( stream.avail_out <= (unsigned)outChunk.size() );
+            out.write( outChunk.data(), (unsigned)outChunk.size() - stream.avail_out );
             if ( out.bad() )
                 return unexpected( "I/O error" );
